@@ -14,9 +14,9 @@
 	"devDependencies": {
 		"@types/common-tags": "^1.8.1",
 		"@types/node": "^16.11.6",
-		"@types/react": "17.0.2",
+		"@types/react": "18.2.38",
 		"@types/react-calendar": "^3.9.0",
-		"@types/react-dom": "17.0.2",
+		"@types/react-dom": "18.2.17",
 		"@typescript-eslint/eslint-plugin": "5.29.0",
 		"@typescript-eslint/parser": "5.29.0",
 		"builtin-modules": "3.3.0",
@@ -30,10 +30,9 @@
 		"@popperjs/core": "^2.11.6",
 		"common-tags": "^1.8.2",
 		"dayjs": "1.11.7",
-		"preact": "10",
-		"react": "npm:@preact/compat@17.0.2",
+		"react": "18.2.0",
 		"react-calendar": "4.0.0",
-		"react-dom": "npm:@preact/compat@17.0.2",
+		"react-dom": "18.2.0",
 		"react-icons": "4.8.0"
 	}
 }
